requirements.txt
PyQt5==5.15.10
requests==2.31.0
Pillow==10.2.0
matplotlib==3.8.2
numpy==1.26.3
openai==1.12.0
colorthief==0.2.1
